+++ demo/main.js
@@ -1,11 +1,14 @@
 import Vue from 'vue';
-import NotificationBus from '@/utils/notifications';
-import RestClient from '@/rest';
-import Girder, { vuetify } from '@';
+import Vuetify from 'vuetify/lib';
+import GirderPlugin, {
+  NotificationBus, RestClient, vuetifyConfig, registerComponents,
+} from '@/';
 
 import App from './App.vue';
 
-Vue.use(Girder);
+Vue.use(GirderPlugin);
+registerComponents();
+
 const girderRest = new RestClient({
   apiRoot: process.env.VUE_APP_API_ROOT,
 });
@@ -14,6 +17,8 @@ const notificationBus = new NotificationBus(girderRest, {
   useEventSource: true,
 });
 
+const vuetify = new Vuetify(vuetifyConfig);
+
 girderRest.fetchUser().then((user) => {
   if (user) {
     notificationBus.connect();
@@ -21,7 +26,7 @@ girderRest.fetchUser().then((user) => {
 
   new Vue({
     vuetify,
-    render: (h) => h(App),
     provide: { girderRest, notificationBus },
+    render: (h) => h(App),
   }).$mount('#app');
 });
