+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@girder/components",
-  "version": "2.2.10",
+  "version": "2.3.0",
   "scripts": {
     "serve": "vue-cli-service serve demo/main.js",
     "build": "vue-cli-service build --target lib --name girder src/index.js",
@@ -12,11 +12,16 @@
   "browser": "dist/girder.umd.min.js",
   "module": "src/index.js",
   "main": "src/index.js",
+  "types": "types/index.d.ts",
   "files": [
     "/dist/",
     "/src/",
+    "/types/",
     "/babel.config.js",
     "/vue.config.js"
+  ],
+  "sideEffects": [
+    "*.css"
   ],
   "dependencies": {
     "@mdi/font": "^5.3.45",
