+++ src/components/Authentication/index.js
@@ -1,3 +1,13 @@
-import Authentication from './Authentication.vue';
+import GirderAuthentication from './Authentication.vue';
+import GirderLogin from './Login.vue';
+import GirderOAuth from './OAuth.vue';
+import GirderRegister from './Register.vue';
 
-export default Authentication;
+export {
+  GirderAuthentication,
+  GirderLogin,
+  GirderOAuth,
+  GirderRegister,
+};
+
+export default GirderAuthentication;
