+++ src/components/Presentation/index.js
@@ -1,0 +1,11 @@
+import GirderDataTable from './DataTable.vue';
+import GirderDetailList from './DetailList.vue';
+import GirderDropzone from './Dropzone.vue';
+import GirderFileUploadList from './FileUploadList.vue';
+
+export {
+  GirderDataTable,
+  GirderDetailList,
+  GirderDropzone,
+  GirderFileUploadList,
+};
