+++ src/components/snippets/index.js
@@ -1,0 +1,4 @@
+import GirderFileManager from './FileManager.vue';
+
+// eslint-disable-next-line import/prefer-default-export
+export { GirderFileManager };
