+++ src/index.js
@@ -1,33 +1,18 @@
 import VueAsyncComputed from 'vue-async-computed';
 import Vuetify from 'vuetify/lib';
-import * as components from './components';
-import * as constants from './constants';
-import RestClient from './rest';
-import * as snippets from './snippets';
-import * as utils from './utils';
 
 /**
  * This installs the Vue plugin for this library. This is exported in the top level module,
  * making it a Vue plugin.
  * @param {Object} Vue the Vue prototype to install the plugin into.
  */
-function install(Vue) {
+export default function install(Vue) {
   Vue.use(Vuetify);
   Vue.use(VueAsyncComputed);
 }
 
-/** The default configured Vuetify instance */
-const vuetify = new Vuetify({ ...utils.vuetifyConfig });
-
-export {
-  components,
-  constants,
-  RestClient,
-  snippets,
-  utils,
-  vuetify,
-};
-
-export default {
-  install,
-};
+export * from './components';
+export * from './constants';
+export { default as RestClient } from './rest';
+export * from './snippets';
+export * from './utils';
