--- src/snippets/index.js
@@ -1,4 +1,0 @@
-import FileManager from './FileManager.vue';
-
-// eslint-disable-next-line import/prefer-default-export
-export { FileManager };
