+++ src/utils/index.js
@@ -1,23 +1,9 @@
 import * as mixins from './mixins';
-import vuetifyConfig from './vuetifyConfig';
-import Upload from './upload';
-import DebounceCounter from './debouncer';
-import {
-  createLocationValidator,
-  getLocationType,
-  isRootLocation,
-  getSingularLocationTypeName,
-} from './locationHelpers';
-import NotificationBus from './notifications';
 
-export {
-  mixins,
-  NotificationBus,
-  DebounceCounter,
-  vuetifyConfig,
-  Upload,
-  createLocationValidator,
-  getLocationType,
-  isRootLocation,
-  getSingularLocationTypeName,
-};
+export { default as vuetifyConfig } from './vuetifyConfig';
+export { default as S3UploadManager } from './s3';
+export { default as UploadManager } from './upload';
+export { default as DebounceCounter } from './debouncer';
+export { default as NotificationBus } from './notifications';
+export * from './locationHelpers';
+export { mixins };
